% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEplotZeros .R
\name{BEplotZeros}
\alias{BEplotZeros}
\title{Change Exploratories plot names without zeros to plot names with zeros
This function adds the sortable plot names to an existing dataset.
Based on a column with Bexis plot names (e.g. AEG1), the function creates a column
with sortable plot names (e.g. AEG01). 
Note: This function does the opposite of the "BEplotNonZeros.R" function}
\usage{
BEplotZeros(dat, column, plotnam = "PlotSTD")
}
\arguments{
\item{dat}{a dataset (data.frame) with at least one column with Bexis plot names (e.g. "AEG1)}

\item{column}{the column name of the column containing the Bexis plot names}

\item{plotnam}{the desired name of the new column with sortable plot names. Default is "PlotSTD"}
}
\value{
the same dataset with an extra column containing the sortable plot names (e.g. "AEG01)
}
\description{
Change Exploratories plot names without zeros to plot names with zeros
This function adds the sortable plot names to an existing dataset.
Based on a column with Bexis plot names (e.g. AEG1), the function creates a column
with sortable plot names (e.g. AEG01). 
Note: This function does the opposite of the "BEplotNonZeros.R" function
}
\examples{
#create a dataset with a plot name column
dat <- data.frame(Plot_name = c("AEG1", "AEG2", "HEW4", "SEG8", "SEW10"), Values=1:5)
dat <- BEplotZeros(dat, "Plot_name", plotnam = "Sorted_plot_name")

}
\author{
Caterina Penone
}
