% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_diversity.R
\name{beta.pair.abund_zerospecies}
\alias{beta.pair.abund_zerospecies}
\title{beta.pair.abund handle plots without species}
\usage{
beta.pair.abund_zerospecies(x, index.family = "bray")
}
\arguments{
\item{x}{a data.set with same requirements as for beta.pair.abund *Note* for the project
BetaDivMultifun, the function `prepare_for_betapair` can be used. For more 
information, please read the documentation for beta.pair.}

\item{index.family}{is "bray" as set by default. Not implemented for "ruzicka"}
}
\value{
same as for function beta.pair.abund A list with three dissimilarity matrices.
For more information, please read the
documentation for beta.pair.
}
\description{
Only minimal description is provided, More info in beta.pair_zerospecies. This function
was generated similarly.
Gradient corresponds to nestedness, balanced to turnover in respect of zeros.
}
