% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaDivMultifun.R, R/clean_datasets.R,
%   R/plotting.R
\docType{package}
\name{BetaDivMultifun}
\alias{BetaDivMultifun}
\alias{BetaDivMultifun-package}
\title{BetaDivMultifun}
\description{
BetaDivMultifun

BetaDivMultifun

BetaDivMultifun
}
