% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{calc_avg_number_of_biotic_abiotic_drivers_of_single_functions}
\alias{calc_avg_number_of_biotic_abiotic_drivers_of_single_functions}
\title{calculate average number of biotic/abiotic/LUI drivers of single functions}
\usage{
calc_avg_number_of_biotic_abiotic_drivers_of_single_functions(
  restab2,
  singleEFnames,
  name_of_summary = "bio_abio_lui"
)
}
\arguments{
\item{restab2}{a data.table created in `plot_GDM.Rmd` creating a summary table of 
all maxspline effects for single functions.}

\item{singleEFnames}{a vector created in `plot_GDM.Rmd` containing the names of 
single functions.}

\item{name_of_summary}{is "bio_abio_lui", no other argument is accepted}
}
\description{
calculate the following average numbers to report from analysis :
Number of biotic/LUI/abiotic drivers of each single function separately.
Subsequently take average of all single functions, in order to get average
number of drivers per function per category.
To answer e.g. the following question : "How many biotic groups do
affect a single function on average?"
Note that all parameters have very specific arguments, which are produced in the
script `plot_GDM.Rmd`.
}
