% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{colComb}
\alias{colComb}
\title{combine columns}
\usage{
colComb(xx, oldcols, newcolname, merge.fun = "mean")
}
\arguments{
\item{xx}{input file}

\item{oldcols}{names of the columns to be combined, given as a vector}

\item{newcolname}{the name which the newly generated, combined column should have}

\item{merge.fun}{the name of the function to be used for column combination. e.g. "mean".}
}
\value{
a merged column in the same dataset as given as input.
}
\description{
This function combines columns. It creates a new column which is the mean (or any given function)
of several columns. The old columns are deleted.
It can be used e.g. to combine data from several years.
}
\details{
This code is written by Eric.
}
