% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{CompareToReferenceDataset}
\alias{CompareToReferenceDataset}
\title{Test if column is identical with reference dataset}
\usage{
CompareToReferenceDataset(tocompareset, refset, name, sortcol = "Plot",
  tolerance = 1e-04)
}
\arguments{
\item{tocompareset}{name of the dataset which is compared (dataset 1). Is a data.table.}

\item{refset}{name of the reference dataset (dataset 2). This dataset can be much larger
than \code{tocompareset}, only the two columns of interest are filtered out. 
Is a data.table.}

\item{name}{names of the columns which will be compared.}

\item{sortcol}{name of the column which is used to sort rows}
}
\value{
Boolean value which indicates wheter the columns are identical or not.
}
\description{
Tests wether the column given is identical with another column from a reference dataset.
The order of rows is not identical in both columns, therefore a simple comparison with
\code{identical()} is not possible. 
The columns compared have the same name "name". 
Both datasets have a column (e.g. with the name "Plot"), by which they can be sorted.
Is used in BetaDivMultifun to test wether the dataset content is identical with 
the Synthesis dataset "Info_data_EP grass functions & services.xlsx", BExIS number
}
