% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_abio_barplot}
\alias{create_abio_barplot}
\title{create plot abio}
\usage{
create_abio_barplot(
  type = c("grouped", "stacked"),
  yvar_name = "maxsplines",
  errorbar = F,
  pylim = c(0, 0.44)
)
}
\arguments{
\item{type}{is in "grouped" (turnover and nestedness are shown as individual
bars) or in "stacked" (turnover and nestedness are stacked on top of each
other). This does not matter for abiotic plot, but is kept in order to make
clear it's the same code for abiotic, above and belowground.}

\item{yvar_name}{character vector defining the column name of y values. 
Defaults to "maxsplines" as created in `create_restab0`.}

\item{errorbar}{logical T or F. If T, error bars (standard errors) are shown.
The name of the errorbar is assumed to be "se_" yvar_name as in `paste0("se_", yvar_name)`
Note that error bars only look good if type is "grouped"}

\item{pylim}{the limits of the y axis. defaults to 0 and 0.44. Note that if errorbars are plotted,
it might be better to set them to ylim = c(-0.22, 0.44)}
}
\description{
code to create plot abio.
}
