% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_bio_belowground_barplot}
\alias{create_bio_belowground_barplot}
\title{create plot bio belowground}
\usage{
create_bio_belowground_barplot()
}
\description{
code to create plot p : bio and belowground
}
