% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_gdm_lineplot}
\alias{create_gdm_lineplot}
\title{create lineplots}
\usage{
create_gdm_lineplot(data, legend = F, ymax = 1, ymin = 0, ribbon = F)
}
\arguments{
\item{data}{data.table, produced by `gdm::isplineExtract` and further
cleaning (as shown in `plot_gdm.Rmd`).}

\item{legend}{logical T or F, indicating if the plot or the legend should be returned.}

\item{ymax}{the maximum y value shown. Defaults to 1.}

\item{ymin}{the minimum y value shown. Defaults to 0.}

\item{ribbon}{logical. T if error bands should be shown, F otherwise.}
}
\value{
a ggplot2 plot element, with lineplots.
}
\description{
create lineplots based on the I splines extracted from a gdm model
with the function `gdm::isplineExtract`.
Used the overview table `ǹicenames.csv` for graphical parameters as
line types and color.
Can plot a legend if legend is set to T, but the legend is recommended
to be produced by the function `get_nice_legend`.
}
\details{
TODO : add vertical grid lines, remove x axis for above plots
}
