% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multifun_thresholds.R
\name{create_gdm_lineplot_predictorwise_thresholds}
\alias{create_gdm_lineplot_predictorwise_thresholds}
\title{create lineplots of all thresholds}
\usage{
create_gdm_lineplot_predictorwise_thresholds(
  data,
  legend = F,
  ribbon = T,
  ribbontype = "se",
  ymax = 1,
  plottitle = ""
)
}
\arguments{
\item{data}{data.table, produced in `GDM_multifun_thresholds.Rmd`).}

\item{legend}{logical T or F, indicating if the plot or the legend should be returned.}

\item{ribbon}{logical. T if error bands should be shown, F otherwise.}

\item{ribbontype}{character, either "se" or "sd". Defines the type of errorbands shown
if ribbon is set T (i.e. error bands are shown)}

\item{ymax}{the maximum y value shown. Defaults to 1.}

\item{plottitle}{a character giving the title of the plot. Defaults to no title.}
}
\value{
a ggplot2 plot element, with lineplots.
}
\description{
Creat lineplots for the weighted average models over all thresholds.
The basic shape and appearance is the same as in GDM lineplots, when
using the function `create_gdm_lineplot`.
}
\details{
The shown lines are either (1) the extracted Isplines from the GDM
models, or (2) the weighted averages over all thresholds. The weighted
averages are calculated at each X value, per predictor, per model type
(turnover / nestedness). The predicted Y values are weighted by their inverse 
standard deviation.

Used the overview table `ǹicenames.csv` for graphical parameters as
line types and color.
Can plot a legend if legend is set to T, but the legend is recommended
to be produced by the function `get_nice_legend`.

TODO : remove x axis for above plots
}
