% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_overview_above_below_abiotic_barplot}
\alias{create_overview_above_below_abiotic_barplot}
\title{create plot Overview : Aboveground - belowground - abiotic}
\usage{
create_overview_above_below_abiotic_barplot(legend = F)
}
\arguments{
\item{legend}{boolean T or F. if T, a plot with legend is created.
The legend can be exctracted with cowplot.}
}
\description{
code to create overview
}
