% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_overview_barplot}
\alias{create_overview_barplot}
\title{Create Overview Barplot}
\usage{
create_overview_barplot(df, legend = F)
}
\arguments{
\item{df}{input dataset for plotting, taken from `create_overviewbar_restab()`.}

\item{legend}{if set to T, only the legend is the output. Defaults to F.}
}
\value{
A barplot of the overviewbars. If `legend=T`, legend extracted with `cowplot::get_legend` which can be plotted
using `cowplot::plot_grid(legend)`.
}
\description{
Creates overview barplots from the data.table outputted by `create_overviewbar_restab()`. 
Plots all overviewbars provided by the input data, in order
to cut bars, please take out the respective lines.
}
