% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_restab}
\alias{create_restab}
\title{Create the table for plotting}
\usage{
create_restab(x = 1)
}
\value{
A data.table ready for plotting
}
\description{
Create the table for plotting
}
\examples{
blabla TODO
}
