% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_restab2}
\alias{create_restab2}
\title{Create the table 2 for plotting}
\usage{
create_restab2(x = 1)
}
\value{
A data.table ready for plotting
}
\description{
Create the table 2 for plotting
}
\examples{
blabla TODO
}
