% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_single_funs_overviewbar_plot}
\alias{create_single_funs_overviewbar_plot}
\title{Create the overview plots for single functions models}
\usage{
create_single_funs_overviewbar_plot(singleF_restab, legend = F, pos = "stack")
}
\arguments{
\item{singleF_restab}{the input table, created with `create_single_funs_overviewbars`}

\item{pos}{defines the bar position, either stack for bars on top of each other (stacked bars)
or "dodge" for bars next to each other}
}
\value{
a ggplot element containing the plot
}
\description{
create overview barplot of single functions, above- belowground and turnover/ nestedness
}
