% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{define_rownames_lui_or_components}
\alias{define_rownames_lui_or_components}
\title{adjust rownames to lui or components}
\usage{
define_rownames_lui_or_components(luiinput)
}
\arguments{
\item{luiinput}{a character vector, either "lui" or "components"}
}
\value{
`plotsequence_abio`, a character vector containing the names for the LUI rows in `restab`,
the table used for plotting.
}
\description{
If LUI is given, delta and sd is only shown for LUI itself.
If components are given, delta F, delta M and delta G  as well as sd F, sd M and
sd G are given.
This function checks which lui input is given and adds the required placeholders
to the vector.
This will make possible the construction of restab accordingly.
}
