% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_diversity.R
\name{find_and_supplement_missing_plots}
\alias{find_and_supplement_missing_plots}
\title{Fill number of plots to 150}
\usage{
find_and_supplement_missing_plots(diversity_dataset, all_plots)
}
\arguments{
\item{diversity_dataset}{a species diversity data.table in the format of the
exploratories synthesis diversity dataset. It contains at least the columns Plot, 
Species and value.}

\item{all_plots}{a character vector containing the complete set of plots from the 
Exploratories. (AEG 1-50, HEG 1-50 and SEG 1-50).}
}
\value{
a data.table consisting of the unchanged diversity_dataset as it has been provided 
as input, rbinded to a minimal data.table containing a random species and the 
missing plots.
}
\description{
Subfunction for add_back_missing_plots_species_combinations
A dataset from soilfungi or bacteria can miss some plots. This happens if there are plots
without any species (e.g. for symbiont.soilfungi in 2011).
The dataset needs to be extended with the missing plots in order to use the function
`add_back_missing_plots_species_combinations`, because this function only adds the missing
plot x species combination, not completely missing plots.
}
