% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{get_nice_legend}
\alias{get_nice_legend}
\title{get legends of plots}
\usage{
get_nice_legend(type)
}
\arguments{
\item{type}{character, is either "biotic" or "abiotic".}
}
\value{
`legend`, legend extracted with `cowplot::get_legend` which can be plotted
using `cowplot::plot_grid(legend)`.
}
\description{
based on the overview table `ǹicenames.csv`, the legend shows colors
for the different predictors.
the legend is split into abiotic and biotic predictors, as there are
16 different colors, but 28 different predictors.
TODO : add the overview bar colors as well, to produce legend for them.
}
