% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathfuns.R
\name{get_random_number_from_truncated_normal_dist}
\alias{get_random_number_from_truncated_normal_dist}
\title{Get random vector from a truncated normal distribution}
\usage{
get_random_number_from_truncated_normal_dist(length_of_vector, mean, max,
  min)
}
\arguments{
\item{length_of_vector}{number, length of returned vector}

\item{mean}{number, the mean of the normal distribution.}

\item{max}{number, the maximum which is assigned to all values which
exceed the limits}

\item{min}{number, the minimum which is assigned to all values which 
are below the given limit.}
}
\value{
a numeric vector of randomly generated numbers
}
\description{
Get a vector of random numbers, drawn from a normal distribution, 
which is truncated (e.g. between 0 and 1).
Has been used to generate the example dataset.
The mean is given, the standard deviation is the mean / 3.
}
