% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_dataset.R
\name{hedo}
\alias{hedo}
\title{show head in two directions}
\usage{
hedo(x, n = 10)
}
\arguments{
\item{x}{Input data frame or matrix.}

\item{n}{Show the first n rows and columns, defaults to 10.}
}
\value{
The first n rows and columns of a data frame or matrix.
}
\description{
To show the first 10 lines of a table, \code{head()} can be used. To show
the first 10 columns as well, \code{head(x[,1:10])} can be used. The 
"\code{hedo()}" function 
shortens this command to \code{hedo(x,10)}. The number of 
rows/columns shown can be changed to any value.
}
\examples{
m <- matrix(sample(seq(0,100),225, replace=T), 15, 150)
m
head(m)
hedo(m)
}
