% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isplineExtract.edit.R
\name{isplineExtract.edit}
\alias{isplineExtract.edit}
\title{isplineExtract.edit}
\usage{
isplineExtract.edit(model)
}
\description{
This is an edit of the function in order to "handle" NULL input
 returns NULL output for NULL input.
}
