% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_environment.R
\name{ksource}
\alias{ksource}
\title{Sourcing a Rmarkdown document}
\usage{
ksource(x, ...)
}
\arguments{
\item{x}{path to the Rmd file}
}
\value{
the output of the sourced file
}
\description{
Function extracted from : https://stackoverflow.com/questions/10966109/how-to-source-r-markdown-file-like-sourcemyfile-r
Sources an Rmd file by creating a temporary R file.
Aim : source Rmd files as .R files (not possible at this moment)
}
\details{
Example use (not run) : ksource("vignettes/prepare_and_run_GDM.Rmd")
}
