% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{loadAndCleanEcosystemFunctions}
\alias{loadAndCleanEcosystemFunctions}
\title{load and clean functions dataset}
\usage{

  loadAndCleanEcosystemFunctions(dataset = "EP grass functions & services.txt",
  path, remove = c("Total.flower.cover", "Charism.butterfly.Abund",
  "Charism.butterfly.SR", "SR.birds", "forage.quality", "daily.temp.range",
  "daily.moisture.range", "P_loss2015", "Soil.C.stock"))
}
\arguments{
\item{dataset}{input file}

\item{path}{path to input file, ending in "/"}

\item{remove}{vector of column names to remove. Defaults to all services
and the two functions "P loss 2015" and "soil C stock".}
}
\value{
a cleaned functions dataframe
}
\description{
This function loads the functions dataset.
It removes all columns mentioned, by default those are all services
and the two functions "P loss 2015" and "soil C stock".
}
\details{
The function automatically calculates shoot Nitrogen stocks and
shoot Phosphorous stocks, both from years 2009 and 2011. Both are calculated
by multiplication of the Biomass measured in the given year and the shoot
Nitrogen/ Phosphorous stock.

The function further takes the means of variables measured in several
years. By default, those are : Nitrogen and Phosphorous stock,
Biomass, Phosphorous loss, Phoshporous shoot, Phoshporous in mycorrhiza

Colcomb is used to calculate the mean level of functions which are
measured in more than one year. Those are: shootPstock, shootNstock,
Biomass, P_loss, Pshoot, Pmic.

This code is written by Eric.
}
