% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_datasets.R
\name{loadAndCleanLUI}
\alias{loadAndCleanLUI}
\title{load and clean LUI dataset}
\usage{
loadAndCleanLUI(dataset = "LUI06to10update20.07.16.txt", path)
}
\arguments{
\item{dataset}{input file}

\item{path}{path to input file, ending in "/"}
}
\value{
a vector with the land-use intensity averages.
}
\description{
This function loads the LUI dataset and calculates the mean of a given
column (e.g. "LUI")
among all given years. The years are given in a column named "Year".
The input data needs to be standardised among years and plots.
It outputs a vector with the land-use intensity averages.
This code is written by Eric.
}
