% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidiversity.R
\name{maxx}
\alias{maxx}
\title{Calculate the mean of top X values}
\usage{
maxx(x, n, ...)
}
\arguments{
\item{x}{a vector?}

\item{n}{Specifies the number of values which are chosen for mean calculation.}
}
\value{
The maximum based on several values (e.g. mean of top 5).
}
\description{
created by Eric Allan
}
\examples{
blabla TODO

}
