% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{panel.cor}
\alias{panel.cor}
\title{Panel cor}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\value{
correlation panel in paris() plot
}
\description{
Function from https://www.rdocumentation.org/packages/graphics/versions/3.6.2/topics/pairs
add correlation panel to `pairs()` plot.
}
