% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{panel.hist}
\alias{panel.hist}
\title{Panel hist}
\usage{
panel.hist(x, ...)
}
\value{
histogram panel in paris() plot
}
\description{
Function from https://www.rdocumentation.org/packages/graphics/versions/3.6.2/topics/pairs
add histogram panel to `pairs()` plot.
}
