% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{paste_gdm_input_path_together}
\alias{paste_gdm_input_path_together}
\title{create GDM input file path}
\usage{
paste_gdm_input_path_together(pathtoout, name)
}
\description{
pastes together the path where to get the gdm input file (.rds) which has been
created before
}
