% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagn_gdm.R
\name{plot_diagn_gdm}
\alias{plot_diagn_gdm}
\title{Plot diagnistic Plot from gdm package}
\usage{
plot_diagn_gdm(x, title = "")
}
\arguments{
\item{x}{gdm model (output from gdm::gdm())}

\item{title}{character string. The title to show on the plot, e.g. the model name.
Defaults to empty.}
}
\value{
the diagnostic plots from GDM model.
}
\description{
This function plots the diagnostic plots from a fitted gdm model.
The package function does the plots, although only together with
all the indivdiual effects, which results in 2+(number of explanatory)
plots.
The function uses 200 samples along the ecological gradient to 
get the line. (similar as used in gdm package.)
}
