% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUncertainty_getsd_GDM.R
\name{plotUncertainty_slim}
\alias{plotUncertainty_slim}
\title{plotUncertainty_slim : Get spline uncertainty}
\usage{
plotUncertainty_slim(
  spTable,
  sampleSites,
  bsIters,
  geo = FALSE,
  splines = NULL,
  knots = NULL,
  splineCol = "blue",
  errCol = "grey80",
  plot.linewidth = 2,
  plot.layout = c(2, 2),
  parallel = FALSE,
  cores = 2,
  save = FALSE,
  fileName = "gdm.plotUncertainy.csv"
)
}
\arguments{
\item{spTable}{see `gdm::plotUncertainty` function description}

\item{sampleSites}{see `gdm::plotUncertainty` function description. 0.3 was used in Gossner 2016.}

\item{bsIters}{see `gdm::plotUncertainty` function description. 100 were used in Gossner 2016.}

\item{geo}{is set to FALSE by default, see `gdm::plotUncertainty` function description}

\item{splines}{is set to NULL by default, see `gdm::plotUncertainty` function description}

\item{knots}{is set to NULL by default, see `gdm::plotUncertainty` function description}

\item{parallel}{is set to FALSE by default, see `gdm::plotUncertainty` function description}

\item{cores}{is set to 2 by default, see `gdm::plotUncertainty` function description}
}
\value{
a data.frame containing "6 values for each predictor, three each for the x and
y values. For each x and y, there are the full model values, the upper and the lower bound."
(taken from the original `gdm::plotUncertainty` function description)
}
\description{
This function is a "slim" version of the `gdm::plotUncertainty` function.
}
\details{
Two changes have been applied to the functio : (1)
The plotting has been disabled, because the function is runned on a cluster
and no plotting output is desired (or can later be done based on the extracted)
values.
(2) The saving option is edited, such that the output of the function is returned
to a variable rather than saved as .csv. This is more appropriate in the
cluster surrounding.

All parameters are the same as in the original function, but a few of them are 
depreciated :
unneeded plotting variables are : `splineCol`, `errCol`, `plot.linewidth`, `plot.layout`
unneeded saving variables are : `save`, `fileName`
Please consult the original description of the function by the `gdm` package authors,
because this is really just a simple edit.
}
\examples{
# read in gdminput from `results_nonpublic.R`.
test <- plotUncertainty_slim(gdminput, sampleSites = 0.3, bsIters = 5, geo = T)

}
