% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_diversity.R
\name{prepare_for_betapair}
\alias{prepare_for_betapair}
\title{Prepare for beta.part}
\usage{
prepare_for_betapair(datatable)
}
\arguments{
\item{datatable}{Is a data.table in which the following 3 columns are present : 
Species, Plot and value. Species describes the taxon whose presence or absence is 
assessed. Plot is the (useful) Name of the site, in the Biodiversity 
Exploratories, there are maximim 150 different Plots possible. Value is either
0 or 1, where 1 means that the Species in the according row is present in the 
according Plot. In case the data table has more columns, this does not matter.}
}
\value{
A data.frame which corresponds to the input requirements of `beta.pair`.
This means, rownames are Plots, Column names are species. And all data in the 
data frame is numeric.
}
\description{
This function prepares the diversity data stored in a data.table for calculation
of betadiversity by `betapart::beta.pair`.
}
