% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathfuns.R
\name{scale01}
\alias{scale01}
\title{Normalisation / scale to 0 and 1}
\usage{
scale01(x)
}
\arguments{
\item{x}{an input vector of values to be scaled.}
}
\value{
a vector of scaled values to be between 0 and 1.
}
\description{
This function normalises a vector to be between 0 and 1. The minimum
is substracted from a sample and the result is divided by the range
of the vector.
NA values are ignored.
\deqn{ \frac{ x - min(x) }{ max(x) - min(x) } = \frac{x - min(x)}{range(x)}}{(x - min(x) / (max(x) - min(x)) )}
}
\examples{
x = sample(-100:100, 50) # create sample
normalized = (x-min(x))/(max(x)-min(x)) # normalized Data
par(mfrow=c(1,2)) # plot both example data and normalized data
plot(x)
plot(normalized) # the shape is not affected, only the range changed

}
