% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_environment.R
\name{setNonpublicVariables}
\alias{setNonpublicVariables}
\title{set nonpublic variables}
\usage{
setNonpublicVariables(accessBexispath = ".", pathtoRAWdata = ".",
  pathtoASSEMBLEDdata = ".")
}
\arguments{
\item{accessBexispath}{the path to the file "accesBexis.R".
This file is available at BExIS and contains code to import data 
directly from R.}

\item{pathtoRAWdata}{path to the data used for this analysis. This
is where the raw data, downloaded from BExIS is stored. It can be
the same folder as pathtoASSEMBLEDdata or not.}

\item{pathtoASSEMBLEDdata}{path to the data used for this analysis. 
This path points to data which is not originally from BExIS but 
has been assembled by someone already. It can be
the same folder as pathtoRAWdata or not.}
}
\value{
a message that the variables have been set
}
\description{
This function sets some global variables which are useful
for a lazy use of this package. Set those variables once and
they will be used as defaults in the package.
}
