% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_overview_above_below_abiotic_barplot}
\alias{create_overview_above_below_abiotic_barplot}
\title{DEPRECIATED create plot Overview : Aboveground - belowground - abiotic
TODO delete this function, use now : `create_overview_barplot` 18.11.22}
\usage{
create_overview_above_below_abiotic_barplot(df, legend = F)
}
\arguments{
\item{df}{data frame to be used for plotting. Is output from `create_overviewbar_restab`.}

\item{legend}{boolean T or F. if T, a plot with legend is created.
The legend can be exctracted with cowplot.1}
}
\description{
code to create overview
}
