% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gdm_FUN.R
\name{create_overview_turnover_nestedness_abiotic_barplot}
\alias{create_overview_turnover_nestedness_abiotic_barplot}
\title{create plot Overview : turnover - newtedness - abiotic}
\usage{
create_overview_turnover_nestedness_abiotic_barplot(legend = F)
}
\arguments{
\item{legend}{boolean, if T, only legend is returned, if F, 
only plot is returned.}
}
\description{
code to create overview
}
