% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_overviewbar_restab}
\alias{create_overviewbar_restab}
\title{Create restab for overview bars}
\usage{
create_overviewbar_restab(restab)
}
\arguments{
\item{restab}{exactly the output from the function `create_restab0()`}
}
\value{
a data.table containing values ready for plotting
}
\description{
calculates scaled averages of effect sizes per category, either (1) above- or belowground
or (2) turnover or nestedness, as well as LUI and abiotic effects. The categories correspond
to the categories given in the table `nicenames` in columns`lui_ground` and `lui_component`.
}
\details{
*Note for developers* : This function was built from `create_restab2()` and `create_restab_3()`
which were deleted (available in git history, date : 22.06.22)
}
