% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_single_funs_overviewbars}
\alias{create_single_funs_overviewbars}
\title{Create the overview table for single functions models}
\usage{
create_single_funs_overviewbars(restab2)
}
\arguments{
\item{restab2}{the input table, created in chunk for single functions heatmap}
}
\value{
A list with 2 data.tables ready for plotting
}
\description{
create overview bars of single functions, above- belowground and turnover/ nestedness
instead of maxsplines as for single models, here we work with all functions. names of all
functions are stored in vector `singleEFnames`
}
\examples{
blabla TODO

}
